requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap4==3.0.1
django-braces==1.14.0
django-decouple==2.1
django-embed-video==1.4.0
django-model-utils==4.1.1
django-registration==2.4.1
django-star-ratings==0.9.1
django-tinymce==3.3.0
django-utils-six==2.0
django-xhtml2pdf==0.0.4
djangorestframework==3.12.4
future==0.18.2
html5lib==1.1
idna==2.10
importlib-metadata==2.1.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
reportlab==3.6.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
sweetify==2.1.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
zipp==3.4.1
